requirements.txt
Jinja2==3.1.4
jinja2-cli==0.8.2
kaleido==0.2.1
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
plotly==5.22.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tenacity==8.4.1
tzdata==2024.1
